refactor(#151): hoist parse_table_info too (simplify follow-up)
Independent review found a larger remaining copy-paste than the helpers I
first hoisted: the PRAGMA table_info parsing tail (~30 lines: the name/type/
notnull/pk index lookups, the row loop building ColumnInfo + primary_key, and
the TableInfo construction) was byte-identical in both adapters, differing only
in how they fetch (columns, rows). Moved it to adapter_common::parse_table_info;
each adapter's table_info now just obtains (columns, rows) and calls it.
cached_table_info stays per-adapter (reads self state). Pure move, no behavior
change.

wasm build + wasm clippy + host clippy -D warnings + fmt clean.
